requirements.txt
# Salesforce Pipeline Automation - Python Dependencies

# Core data processing
pandas>=1.5.0
openpyxl>=3.0.0

# GUI Framework
Eel>=0.14.0

# Environment variables
python-dotenv>=0.19.0

# Built-in modules (no installation needed):
# - imaplib (email receiving)
# - smtplib (email sending) 
# - email (email parsing)
# - tkinter (file dialogs)
# - os, sys, pathlib (system operations)
# - logging (logging)
# - subprocess (running external processes)
# - datetime (timestamps)
# - shutil (file operations)

# Optional for development/testing
# pytest>=7.0.0
# black>=22.0.0